requirements.txt
openpyxl>=3.1.0
python-docx>=1.0.0
streamlit>=1.32.0
weasyprint>=61.0
